Added a README file to the 3rd-party directory explaining the purpose of these third party libraries. For two of the libraries I have no idea what they are used for (tr and gl2ps). From the line coverage analysis they also do not appear to be used. Can somebody explain these, or better remove them.
Furthermore, a next round of Windows 64 bit warnings have been
removed and more dead code has been eliminated. It appears
to me that only a very small fraction of the Windows warnings
are now in the ATerm library. Moreover, coverage of the remaining
portion of ATerm code is high. 



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/trunk@8513 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ 3rd-party/aterm/include/aterm/afun.h
@@ -27,8 +27,8 @@ typedef struct _SymEntry
   struct _SymEntry *next;
   AFun  id;
   char   *name;
-  int     count;  /* used in bafio.c */
-  int     index;  /* used in bafio.c */
+  size_t count;  /* used in bafio.c */
+  size_t index;  /* used in bafio.c */
 } *SymEntry;
 
 /* defined on SymEntry */
