+++ 3rd-party/aterm/include/aterm/byteencoding.h
@@ -8,12 +8,12 @@ extern "C"
 {
 #endif /* __cplusplus */
 
-int BEserializeMultiByteInt(int i, char *c);
+size_t BEserializeMultiByteInt(int i, char *c);
 
 /* void BEserializeDouble(double d, char *c); */
 
 
-int BEdeserializeMultiByteInt(char *c, size_t *i);
+size_t BEdeserializeMultiByteInt(char *c, size_t *i);
 
 /* double BEdeserializeDouble(char *c); */
