+++ 3rd-party/aterm/source/byteio.c
@@ -149,7 +149,7 @@ void init_file_reader(byte_reader *reader, FILE *file)
 /*}}}  */
 /*{{{  void init_string_reader(byte_reader *reader, const unsigned char *buf, int max_size) */
 
-void init_string_reader(byte_reader *reader, const unsigned char *buf, int max_size)
+void init_string_reader(byte_reader *reader, const unsigned char *buf, size_t max_size)
 {
   reader->type = STRING_READER;
   reader->bytes_read = 0;
